:art: refactor shared imports
