+++ app/facades/acapy_ledger.py
@@ -10,8 +10,8 @@ from aries_cloudcontroller import (
     TxnOrRegisterLedgerNymResponse,
 )
 
-from app.config.log_config import get_logger
 from app.exceptions.cloud_api_error import CloudApiException
+from shared.log_config import get_logger
 
 logger = get_logger(__name__)
