+++ app/facades/revocation_registry.py
@@ -13,10 +13,10 @@ from aries_cloudcontroller import (
     TxnOrRevRegResult,
 )
 
-from app.config.log_config import get_logger
 from app.dependencies.acapy_clients import get_governance_controller
 from app.event_handling.sse_listener import SseListener
 from app.exceptions.cloud_api_error import CloudApiException
+from shared.log_config import get_logger
 
 logger = get_logger(__name__)
