+++ app/generic/issuer/facades/acapy_issuer_v2.py
@@ -13,12 +13,12 @@ from aries_cloudcontroller import (
 )
 from aries_cloudcontroller.model.v20_cred_store_request import V20CredStoreRequest
 
-from app.config.log_config import get_logger
 from app.exceptions.cloud_api_error import CloudApiException
 from app.generic.issuer.facades.acapy_issuer import Issuer
 from app.generic.issuer.models import Credential, CredentialNoConnection
 from app.util.credentials import cred_id_no_version
-from shared import CredentialExchange, credential_record_to_model_v2
+from shared.log_config import get_logger
+from shared.models import CredentialExchange, credential_record_to_model_v2
 
 logger = get_logger(__name__)
