+++ app/generic/messaging.py
@@ -3,9 +3,9 @@ from aries_cloudcontroller.model.ping_request_response import PingRequestRespons
 from fastapi import APIRouter, Depends
 from pydantic import BaseModel
 
-from app.config.log_config import get_logger
 from app.dependencies.acapy_clients import client_from_auth
 from app.dependencies.auth import AcaPyAuth, acapy_auth
+from shared.log_config import get_logger
 
 logger = get_logger(__name__)
