+++ app/generic/trust_registry.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 
 import app.facades.trust_registry as trust_registry_facade
-from app.config.log_config import get_logger
+from shared.log_config import get_logger
 
 logger = get_logger(__name__)
