+++ app/generic/wallet/wallet.py
@@ -3,12 +3,12 @@ from typing import List
 from aries_cloudcontroller import DID, DIDEndpoint, DIDEndpointWithType
 from fastapi import APIRouter, Depends
 
-from app.config.log_config import get_logger
 from app.dependencies.acapy_clients import client_from_auth
 from app.dependencies.auth import AcaPyAuth, acapy_auth
 from app.exceptions.cloud_api_error import CloudApiException
 from app.facades import acapy_wallet
 from app.generic.wallet.models import SetDidEndpointRequest
+from shared.log_config import get_logger
 
 logger = get_logger(__name__)
