+++ app/main.py
@@ -11,7 +11,7 @@ from fastapi.exceptions import HTTPException
 from fastapi.responses import JSONResponse
 
 from app.admin.tenants import tenants
-from app.config.log_config import get_logger
+from shared.log_config import get_logger
 from app.event_handling.webhooks import Webhooks
 from app.exceptions.cloud_api_error import CloudApiException
 from app.generic import definitions, messaging, sse, trust_registry, webhooks
