+++ app/tests/e2e/test_fixtures.py
@@ -15,7 +15,8 @@ from app.tests.util.ecosystem_connections import FaberAliceConnect
 from app.tests.util.string import random_version
 from app.tests.util.trust_registry import register_issuer
 from app.tests.util.webhooks import check_webhook_state
-from shared import CredentialExchange, RichAsyncClient
+from shared import RichAsyncClient
+from shared.models import CredentialExchange
 
 CREDENTIALS_BASE_PATH = router.prefix + "/credentials"
