+++ app/tests/e2e/test_issuer.py
@@ -8,7 +8,8 @@ from app.tests.e2e.test_fixtures import CREDENTIALS_BASE_PATH
 from app.tests.util.ecosystem_connections import FaberAliceConnect
 from app.tests.util.webhooks import check_webhook_state, get_wallet_id_from_async_client
 from app.util.credentials import cred_id_no_version
-from shared import CredentialExchange, RichAsyncClient
+from shared import RichAsyncClient
+from shared.models import CredentialExchange
 
 
 @pytest.mark.anyio
