+++ app/tests/e2e/test_verifier.py
@@ -18,11 +18,11 @@ from app.generic.verifier.models import (
 from app.tests.util.ecosystem_connections import AcmeAliceConnect
 from app.tests.util.webhooks import check_webhook_state, get_wallet_id_from_async_client
 from app.tests.verifier.utils import indy_proof_request
-from shared import (
+from shared import RichAsyncClient
+from shared.models import (
     CredentialExchange,
     PresentationExchange,
     PresentProofProtocolVersion,
-    RichAsyncClient,
 )
 
 VERIFIER_BASE_PATH = "/generic/verifier"
