+++ app/tests/e2e/test_webhook.py
@@ -3,7 +3,8 @@ from fastapi import HTTPException
 
 from app.generic.webhooks import router
 from app.tests.util.ecosystem_connections import BobAliceConnect
-from shared import Connection, RichAsyncClient
+from shared import RichAsyncClient
+from shared.models import Connection
 
 WALLET_BASE_PATH = router.prefix
