+++ app/tests/verifier/test_verifier.py
@@ -13,7 +13,7 @@ from app.generic.verifier.facades.acapy_verifier_v1 import VerifierV1
 from app.generic.verifier.facades.acapy_verifier_v2 import VerifierV2
 from app.tests.util.mock import to_async
 from app.tests.verifier.utils import indy_pres_spec, indy_proof_request
-from shared import PresentationExchange, PresentProofProtocolVersion
+from shared.models import PresentationExchange, PresentProofProtocolVersion
 from shared.util.mock_agent_controller import MockContextManagedController
 
 presentation_exchange_record_1 = PresentationExchange(
