+++ endorser/main.py
@@ -1,7 +1,7 @@
 import asyncio
 
-from endorser.config.log_config import get_logger
 from endorser.endorser_processor import listen_endorsement_events
+from shared.log_config import get_logger
 
 logger = get_logger(__name__)
