+++ shared/__init__.py
@@ -1,9 +1,5 @@
 # flake8: noqa
 from shared.constants import *
-from shared.models.conversion import *
-from shared.models.protocol import *
-from shared.models.topics import *
-from shared.models.topics.base import *
 from shared.util.api_router import APIRouter
 from shared.util.rich_async_client import RichAsyncClient
 from shared.util.rich_parsing import parse_with_error_handling
