+++ shared/models/__init__.py
@@ -1,0 +1,4 @@
+from shared.models.conversion import *
+from shared.models.protocol import *
+from shared.models.topics import *
+from shared.models.topics.base import *
