+++ trustregistry/main.py
@@ -3,8 +3,8 @@ import os
 from fastapi import Depends, FastAPI
 from sqlalchemy.orm import Session
 
+from shared.log_config import get_logger
 from trustregistry import crud, models
-from trustregistry.config.log_config import get_logger
 from trustregistry.database import engine
 from trustregistry.db import get_db
 from trustregistry.registry import registry_actors, registry_schemas
