+++ trustregistry/registry/registry_schemas.py
@@ -5,8 +5,8 @@ from fastapi.params import Depends
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 
+from shared.log_config import get_logger
 from trustregistry import crud
-from trustregistry.config.log_config import get_logger
 from trustregistry.db import get_db
 from trustregistry.schemas import Schema
