+++ webhooks/routers/webhooks.py
@@ -4,7 +4,7 @@ from dependency_injector.wiring import Provide, inject
 from fastapi import Depends
 
 from shared import APIRouter, TopicItem
-from webhooks.config.log_config import get_logger
+from shared.log_config import get_logger
 from webhooks.dependencies.container import Container
 from webhooks.dependencies.service import Service
